+++ database/migrations/2018_01_31_172203_create_links_table.php
@@ -20,7 +20,7 @@ class CreateLinksTable extends Migration
             $table->string('parent',45)->nullable();
             $table->string('desc')->nullable();
             $table->string('dest')->nullable();
-            $table->smallInteger('media_id')->nullable();
+            $table->unsignedInteger('media_id')->nullable();
             $table->string('icon',45)->nullable();
             $table->timestamps();
